refactor(tui): split status_modifier tests to keep render_tests.rs under cap
`render_tests.rs` reached 480 lines after the #539 additions, breaking the
400-LOC file-size lint. Move the 14 pure-function `status_modifier_*` tests
into a sibling `status_modifier_tests.rs` (wired via a second `#[cfg(test)]
#[path = ...]` mod under render.rs). render_tests.rs returns to 368 LOC.

The role-color-survives tests stay in render_tests.rs because they share
the buffer-rendering helpers (`make_session_with`, `render_buffer`,
`buffer_has_color`).
